+++ src/commands/env/log.ts
@@ -5,26 +5,26 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import herokuColor from '@heroku-cli/color';
-import * as Heroku from '@heroku-cli/schema';
-import { Errors, Flags } from '@oclif/core';
 import { cli } from 'cli-ux';
+import * as Heroku from '@heroku-cli/schema';
 import { Messages } from '@salesforce/core';
-import { FunctionsFlagBuilder } from '../../../lib/flags';
-import Command from '../../../lib/base';
-import { resolveAppNameForEnvironment } from '../../../lib/utils';
+import { Errors, Flags } from '@oclif/core';
+import { FunctionsFlagBuilder } from '../../lib/flags';
+import Command from '../../lib/base';
+import { resolveAppNameForEnvironment } from '../../lib/utils';
+import * as logStreamUtils from '../../lib/log-stream-utils';
 
 Messages.importMessagesDirectory(__dirname);
-const messages = Messages.loadMessages('@salesforce/plugin-functions', 'env.logdrain.add');
+const messages = Messages.loadMessages('@salesforce/plugin-functions', 'env.log');
 
-export default class LogDrainAdd extends Command {
+export default class Log extends Command {
   static summary = messages.getMessage('summary');
-
-  static description = messages.getMessage('description');
 
   static examples = messages.getMessages('examples');
 
   static flags = {
     'target-compute': FunctionsFlagBuilder.environment({
+      description: messages.getMessage('flags.target-compute.summary'),
       exclusive: ['environment'],
     }),
     environment: FunctionsFlagBuilder.environment({
@@ -32,24 +32,17 @@ export default class LogDrainAdd extends Command {
       exclusive: ['target-compute'],
       hidden: true,
     }),
-    'drain-url': Flags.string({
-      exclusive: ['url'],
-      char: 'l',
-      description: messages.getMessage('flags.drain-url.summary'),
-    }),
-    url: Flags.string({
-      exclusive: ['drain-url'],
-      char: 'u',
-      description: messages.getMessage('flags.drain-url.summary'),
-      hidden: true,
+    num: Flags.integer({
+      char: 'n',
+      description: messages.getMessage('flags.num.summary'),
     }),
   };
 
   async run() {
-    const { flags } = await this.parse(LogDrainAdd);
+    const { flags } = await this.parse(Log);
     // We support both versions of the flag here for the sake of backward compat
     const targetCompute = flags['target-compute'] ?? flags.environment;
-    const url = flags['drain-url'] ?? flags.url;
+    const logLines = flags.num ?? 100;
 
     if (!targetCompute) {
       throw new Errors.CLIError(
@@ -59,32 +52,26 @@ export default class LogDrainAdd extends Command {
       );
     }
 
-    if (!url) {
-      throw new Errors.CLIError(
-        `Missing required flag:
-       -u, --drain-url DRAIN-URL  ${herokuColor.dim('Endpoint that will receive sent logs.')}
-       See more help with --help`
-      );
-    }
-
     if (flags.environment) {
       cli.warn(messages.getMessage('flags.environment.deprecation'));
     }
 
-    if (flags.url) {
-      cli.warn(messages.getMessage('flags.url.deprecation'));
-    }
-
     const appName = await resolveAppNameForEnvironment(targetCompute);
 
-    cli.action.start(`Creating drain for environment ${herokuColor.app(targetCompute)}`);
-
-    await this.client.post<Heroku.LogDrain>(`/apps/${appName}/log-drains`, {
+    const response = await this.client.post<Heroku.LogSession>(`/apps/${appName}/log-sessions`, {
       data: {
-        url,
+        tail: false,
+        lines: logLines,
       },
     });
 
+    const logURL = response.data.logplex_url;
+
+    if (logURL) {
+      await logStreamUtils.readLogs(logURL, false);
+    } else {
+      this.error("Couldn't retreive logs");
+    }
     cli.action.stop();
   }
 }