+++ src/commands/env/logdrain/remove.ts
@@ -68,11 +68,11 @@ export default class LogDrainRemove extends Command {
     }
 
     if (flags.environment) {
-      this.warn(messages.getMessage('flags.environment.deprecation'));
+      cli.warn(messages.getMessage('flags.environment.deprecation'));
     }
 
     if (flags.url) {
-      this.warn(messages.getMessage('flags.url.deprecation'));
+      cli.warn(messages.getMessage('flags.url.deprecation'));
     }
 
     const appName = await resolveAppNameForEnvironment(targetCompute);